wip: bug while storing trie state
